pyproject.toml
[tool.poetry]
name = "backend"
version = "0.1.0"
description = ""
authors = ["ANGELINE WANG <02ANGELINEWANG@GMAIL.COM>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
django = "^4.1.3"
mypy = "^0.990"
autopep8 = "^2.0.0"
pillow = "^9.3.0"
psycopg2 = "^2.9.5"
djangorestframework = "^3.14.0"
jwt = "^1.3.1"
pyjwt = "^2.6.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
